requirements.txt
streamlit
pandas
prophet
matplotlib
transformers
langchain-openai
python-dotenv
numpy
torch 
statsmodels
opencv-python
